validate edit form
