+++ src/components/modal/CreateNoteModal/CreateNoteModal.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Button, Modal, ModalBody, ModalFooter, ModalHeader } from 'reactstrap';
-import NoteForm from '../NoteForm/NoteForm';
+import NoteForm from '../../form/NoteForm/NoteForm';
 
 interface CreateNoteModalProps {
     isOpen: boolean;
